add tamago/amd64 support, microvm target (#45)
* WiP tamago/amd64 support

* add CPU initialization functions

* tidying

* add amd64 memory initialization

* tidying

* tidying

* initial amd64 cpuinit (LM with 2MB map)

* working amd64 long mode

* enable SSE

* add basic COM1 support

* increase RAM size

* tidying

* halt and catch fire

* package rename riscv -> riscv64 for consistency with GOARCH

* working 16550A UART driver

* triple-fault on microvm exit

* map page entries to cover ramStart+ramSize

* set stack offset

* add IsSet function

* amd64 TSC support

* microvm kvmclock support

* microvm MC146818A RTC driver

* tidying

* tidying

* typo fix

* typo fix

* URL update

* documentation update

* tidying

* tidying

* tidying

* fix and simplify amd64 TLBs, reserve I/O area

* tidying

* initial import of WiP VirtIO driver

* tidying

* move VirtIO device driver outside tamago

* update qemu settings

* tidying

* add IsSet64 function

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* improve VirtIO driver

* tidying

* bump release

* tidying

+++ soc/nxp/imx6ul/timer.go
@@ -33,5 +33,5 @@ func initTimers() {
 
 //go:linkname nanotime1 runtime.nanotime1
 func nanotime1() int64 {
-	return int64(ARM.TimerFn()*ARM.TimerMultiplier + ARM.TimerOffset)
+	return ARM.TimerFn()*ARM.TimerMultiplier + ARM.TimerOffset
 }
